% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-driver.R
\docType{methods}
\name{dbConnect,Pool-method}
\alias{dbConnect,Pool-method}
\title{Fetch a connection from the pool}
\usage{
\S4method{dbConnect}{Pool}(drv, ...)
}
\arguments{
\item{drv}{A pool object.}

\item{...}{Not in use.}
}
\description{
Fetch a DBIConnection object from the pool. This method should
be used with care, since this puts the onus of releasing the
connection on you (the user). While this means that you still
enjoy the performance benefits associated with having a pool,
you lose the benefits of automatic connection management, since
you are now yourself responsible for releasing the connection
(using \code{release(conn)}) at the appropriate time.
}

