% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-pool.R
\docType{methods}
\name{dbDisconnect,DBIConnection-method}
\alias{dbDisconnect,DBIConnection-method}
\title{Release a connection back to the pool}
\usage{
\S4method{dbDisconnect}{DBIConnection}(conn, ...)
}
\description{
A wrapper around \code{release(conn)}. Unlike its namesake in DBI,
in the context of the \code{pool} package, \code{dbDisconnect} does
not close the connection and free the associated resources. Rather,
it simply releases the connection back to pool, which then decides
whether to keep it around or actually destroy it.
}

