% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-connection-sql.R
\docType{methods}
\name{DBI-connection-sql}
\alias{DBI-connection-sql}
\alias{sqlAppendTable,Pool-method}
\alias{sqlAppendTableTemplate}
\alias{sqlCreateTable,Pool-method}
\alias{sqlData,Pool-method}
\title{Creating and manipulating SQL tables.}
\usage{
\S4method{sqlData}{Pool}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{Pool}(con, table, fields, row.names = NA,
  temporary = FALSE, ...)

\S4method{sqlAppendTable}{Pool}(con, table, values, row.names = NA, ...)

sqlAppendTableTemplate(obj, table, values, row.names = NA, prefix = "?",
  ...)
}
\description{
Pool object wrappers around DBIConnection methods that deal
with the creation and manipulation of SQL tables. See
\code{\link[DBI]{sqlData}}, \code{\link[DBI]{sqlCreateTable}}
and \code{\link[DBI]{sqlAppendTable}} for the original
documentation.
}

