% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-wrap.R
\name{DBI-wrap}
\alias{DBI-wrap}
\alias{dbDataType,Pool-method}
\alias{dbGetQuery,Pool,ANY-method}
\alias{dbExecute,Pool,ANY-method}
\alias{dbListFields,Pool,ANY-method}
\alias{dbListTables,Pool-method}
\alias{dbListObjects,Pool-method}
\alias{dbReadTable,Pool,ANY-method}
\alias{dbWriteTable,Pool,ANY-method}
\alias{dbCreateTable,Pool-method}
\alias{dbAppendTable,Pool-method}
\alias{dbExistsTable,Pool,ANY-method}
\alias{dbRemoveTable,Pool,ANY-method}
\alias{dbIsReadOnly,Pool-method}
\alias{make.db.names,Pool,ANY-method}
\alias{isSQLKeyword,Pool,ANY-method}
\alias{SQLKeywords,Pool-method}
\alias{sqlData,Pool-method}
\alias{sqlCreateTable,Pool-method}
\alias{sqlAppendTable,Pool-method}
\alias{sqlInterpolate,Pool-method}
\alias{sqlParseVariables,Pool-method}
\alias{dbQuoteIdentifier,Pool,ANY-method}
\alias{dbUnquoteIdentifier,Pool-method}
\alias{dbQuoteLiteral,Pool-method}
\alias{dbQuoteString,Pool,ANY-method}
\title{Custom DBI methods for Pool}
\usage{
\S4method{dbDataType}{Pool}(dbObj, obj, ...)

\S4method{dbGetQuery}{Pool,ANY}(conn, statement, ...)

\S4method{dbExecute}{Pool,ANY}(conn, statement, ...)

\S4method{dbListFields}{Pool,ANY}(conn, name, ...)

\S4method{dbListTables}{Pool}(conn, ...)

\S4method{dbListObjects}{Pool}(conn, prefix = NULL, ...)

\S4method{dbReadTable}{Pool,ANY}(conn, name, ...)

\S4method{dbWriteTable}{Pool,ANY}(conn, name, value, ...)

\S4method{dbCreateTable}{Pool}(conn, name, fields, ..., row.names = NULL, temporary = FALSE)

\S4method{dbAppendTable}{Pool}(conn, name, value, ..., row.names = NULL)

\S4method{dbExistsTable}{Pool,ANY}(conn, name, ...)

\S4method{dbRemoveTable}{Pool,ANY}(conn, name, ...)

\S4method{dbIsReadOnly}{Pool}(dbObj, ...)

\S4method{make.db.names}{Pool,ANY}(
  dbObj,
  snames,
  keywords = .SQL92Keywords,
  unique = TRUE,
  allow.keywords = TRUE,
  ...
)

\S4method{isSQLKeyword}{Pool,ANY}(
  dbObj,
  name,
  keywords = .SQL92Keywords,
  case = c("lower", "upper", "any")[3],
  ...
)

\S4method{SQLKeywords}{Pool}(dbObj, ...)

\S4method{sqlData}{Pool}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{Pool}(con, table, fields, row.names = NA, temporary = FALSE, ...)

\S4method{sqlAppendTable}{Pool}(con, table, values, row.names = NA, ...)

\S4method{sqlInterpolate}{Pool}(conn, sql, ..., .dots = list())

\S4method{sqlParseVariables}{Pool}(conn, sql, ...)

\S4method{dbQuoteIdentifier}{Pool,ANY}(conn, x, ...)

\S4method{dbUnquoteIdentifier}{Pool}(conn, x, ...)

\S4method{dbQuoteLiteral}{Pool}(conn, x, ...)

\S4method{dbQuoteString}{Pool,ANY}(conn, x, ...)
}
\description{
Wrappers around all DBI methods that check out a connection
(with \code{\link[=poolCheckout]{poolCheckout()}}), perform the operation, and the return the connection
to the pool (with \code{\link[=poolReturn]{poolReturn()}}).
}
\keyword{internal}
