% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-driver.R
\docType{methods}
\name{dbListConnections,Pool-method}
\alias{dbListConnections,Pool-method}
\title{List currently open connections.}
\usage{
\S4method{dbListConnections}{Pool}(drv, ...)
}
\description{
Pool object wrapper around the original DBI method for
DBIDriver objects. See \code{\link[DBI]{dbListConnections}} for
the original documentation.
}

