% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-connection-transaction.R
\docType{methods}
\name{dbWithTransaction,Pool-method}
\alias{dbWithTransaction}
\alias{dbWithTransaction,Pool-method}
\title{Hello}
\usage{
\S4method{dbWithTransaction}{Pool}(conn, code)
}
\arguments{
\item{conn}{param}

\item{code}{param}
}
\description{
Hello
}

