% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demoDb}
\alias{demoDb}
\title{Create a demo SQLite database}
\usage{
demoDb()
}
\description{
This function creates a temporary SQLite database for demonstration purposes.
It populates the database with two tables: 'mtcars' and 'faithful'.
}
\keyword{internal}
