% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object}
\alias{object}
\alias{onActivate}
\alias{onDestroy}
\alias{onPassivate}
\alias{onValidate}
\title{Pooled object methods.}
\usage{
onActivate(object)

onPassivate(object)

onDestroy(object, envir)

onValidate(object)
}
\description{
For backend authors only. Authors should implement all of these,
which are then called in the within the Pool class methods. These
should not be called directly either by backend authors or by the
end users.
}

