% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool-package.R
\docType{package}
\name{pool-package}
\alias{pool}
\alias{pool-package}
\title{pool: Object Pooling}
\description{
Enables the creation of object pools, which make it less computationally expensive to fetch a new object. Currently the only supported pooled objects are 'DBI' connections.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/pool}
  \item \url{http://rstudio.github.io/pool/}
  \item Report bugs at \url{https://github.com/rstudio/pool/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Joe Cheng \email{joe@posit.co}
  \item Barbara Borges
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
