% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{methods}
\name{poolCreate,DBIConnectionFactory-method}
\alias{poolCreate,DBIConnectionFactory-method}
\title{Given a DBIConnectionFactory object, creates a pool}
\usage{
\S4method{poolCreate}{DBIConnectionFactory}(src, minSize = 3, maxSize = Inf,
  ...)
}
\arguments{
\item{src}{DBIConnectionFactory object.}

\item{minSize}{An optional number specifying the minimum
number of objects that the pool should have at all times.}

\item{maxSize}{An optional number specifying the maximum
number of objects that the pool may have at any time.}

\item{...}{Authorization arguments needed by the DBMS instance;
these typically include \code{user}, \code{password}, \code{dbname},
\code{host}, \code{port}, etc.  For details see the appropriate
\code{DBIDriver}.}
}
\description{
Given a DBIConnectionFactory object, creates a pool
}

