% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI.R
\docType{methods}
\name{poolCreate,DBIDriver-method}
\alias{poolCreate,DBIDriver-method}
\title{Create a DBI Database Connection Pool.}
\usage{
\S4method{poolCreate}{DBIDriver}(src, minSize = 3, maxSize = Inf, ...)
}
\arguments{
\item{src}{An object that inherits from \code{\link[DBI]{DBIDriver-class}},
or a character string specifying the name of DBMS driver, e.g.,
"RSQLite", "RMySQL", "RPostgreSQL", or an existing
\code{\linkS4class{DBIConnection}} object (in order to clone an
existing connection).}

\item{minSize}{An optional number specifying the minimum
number of objects that the pool should have at all times.}

\item{maxSize}{An optional number specifying the maximum
number of objects that the pool may have at any time.}

\item{...}{Authorization arguments needed by the DBMS instance;
these typically include \code{user}, \code{password}, \code{dbname},
\code{host}, \code{port}, etc.  For details see the appropriate
\code{DBIDriver}.}
}
\description{
Creates a DBI database connection pool. Check the documentation
of \code{\link{poolCreate}} for a generic overview of the this
function and the Pool object. The main thing to point out is
that, for DBI database connection pooling, the \code{src}
argument should always be a \code{\link[DBI]{DBIDriver-class}},
and it should always be accompanied by the required authorization
arguments (see the example below).
}
\examples{
\dontrun{
pool <- poolCreate(
  drv = MySQL(),
  dbname = "shinydemo",
  host = "shiny-demo.csa7qlmguqrf.us-east-1.rds.amazonaws.com",
  username = "guest",
  password = "guest"
)

dbGetQuery(pool, "SELECT * from City WHERE ID < 5")
#>   ID           Name CountryCode      District Population
#> 1  1          Kabul         AFG         Kabol    1780000
#> 2  2       Qandahar         AFG      Qandahar     237500
#> 3  3          Herat         AFG         Herat     186800
#> 4  4 Mazar-e-Sharif         AFG         Balkh     127800

pool$close()
}
}

