% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool-methods.R
\name{poolReturn}
\alias{poolReturn}
\alias{poolReturn,ANY-method}
\title{Returns an object back to the pool.}
\usage{
poolReturn(object)
}
\arguments{
\item{object}{A pooled object.}
}
\description{
Should be called by the end user if they previously fetched
an object directly using \code{object <- poolCheckout(pool)}
and are now done with said object.
}

