% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-connection-quote.R
\docType{methods}
\name{DBI-connection-quote}
\alias{DBI-connection-quote}
\alias{dbQuoteIdentifier,Pool,character-method}
\alias{dbQuoteIdentifier,Pool,SQL-method}
\alias{dbQuoteIdentifier,Pool,Table-method}
\alias{dbQuoteString,Pool,character-method}
\alias{dbQuoteString,Pool,SQL-method}
\title{SQL quoting.}
\usage{
\S4method{dbQuoteIdentifier}{Pool,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{Pool,SQL}(conn, x, ...)

\S4method{dbQuoteIdentifier}{Pool,Table}(conn, x, ...)

\S4method{dbQuoteString}{Pool,character}(conn, x, ...)

\S4method{dbQuoteString}{Pool,SQL}(conn, x, ...)
}
\arguments{
\item{conn, x, ...}{See \code{\link[DBI]{SQL}}.}
}
\description{
Pool object wrappers around DBIConnection methods that deal
with SQL escaping needs. See \code{\link[DBI]{SQL}} for the
original documentation.
}
