% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI-connection-sql.R
\name{DBI-connection-sql}
\alias{DBI-connection-sql}
\alias{sqlData,Pool-method}
\alias{sqlCreateTable,Pool-method}
\alias{sqlAppendTable,Pool-method}
\title{Creating and manipulating SQL tables.}
\usage{
\S4method{sqlData}{Pool}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{Pool}(con, table, fields, row.names = NA, temporary = FALSE, ...)

\S4method{sqlAppendTable}{Pool}(con, table, values, row.names = NA, ...)
}
\arguments{
\item{con, value, row.names, ...}{See \code{\link[DBI]{sqlData}}.}

\item{table, fields, temporary}{See \code{\link[DBI]{sqlCreateTable}}.}

\item{values}{See \code{\link[DBI]{sqlAppendTable}}.}
}
\description{
Pool object wrappers around DBIConnection methods that deal
with the creation and manipulation of SQL tables. See
\code{\link[DBI]{sqlData}}, \code{\link[DBI]{sqlCreateTable}}
and \code{\link[DBI]{sqlAppendTable}} for the original
documentation.
}
