% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R, R/DBI-pool.R
\name{object}
\alias{object}
\alias{onActivate}
\alias{onPassivate}
\alias{onDestroy}
\alias{onValidate}
\alias{onActivate,ANY-method}
\alias{onPassivate,ANY-method}
\alias{onDestroy,ANY-method}
\alias{onValidate,ANY-method}
\alias{onPassivate,DBIConnection-method}
\alias{onDestroy,DBIConnection-method}
\alias{onValidate,DBIConnection-method}
\title{Pooled object methods.}
\usage{
onActivate(object)

onPassivate(object)

onDestroy(object)

onValidate(object, query)

\S4method{onActivate}{ANY}(object)

\S4method{onPassivate}{ANY}(object)

\S4method{onDestroy}{ANY}(object)

\S4method{onValidate}{ANY}(object, query)

\S4method{onPassivate}{DBIConnection}(object)

\S4method{onDestroy}{DBIConnection}(object)

\S4method{onValidate}{DBIConnection}(object)
}
\arguments{
\item{object}{A pooled object.}

\item{query}{A simple query that can be used to verify that
the \code{object} functions as expected.}
}
\description{
For backend authors only. Authors should implement all of these,
which are then called by the Pool class methods. These should
not be called directly either by backend authors or by the end
users.
}
\keyword{internal}
