% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{pool-package}
\alias{pool}
\alias{pool-package}
\title{pool: Object Pooling}
\description{
Enables the creation of object pools, which make it
    less computationally expensive to fetch a new object. Currently the
    only supported pooled objects are 'DBI' connections.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/pool}
  \item Report bugs at \url{https://github.com/rstudio/pool/issues}
}

}
\author{
\strong{Maintainer}: Joe Cheng \email{joe@rstudio.com}

Authors:
\itemize{
  \item Barbara Borges \email{barb.b.ribeiro@gmail.com}
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
\keyword{internal}
