% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\docType{package}
\name{pool}
\alias{pool}
\alias{pool-package}
\title{Connection Pooling for Databases in R}
\description{
Creates connection pools for various ypes of databases in R to
make it less computationally expensive to get a connection.
}

