% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool-methods.R
\name{poolCheckout}
\alias{poolCheckout}
\alias{poolCheckout,Pool-method}
\title{Checks out an object from the pool.}
\usage{
poolCheckout(pool)
}
\arguments{
\item{pool}{The pool to get the object from.}
}
\description{
Should be called by the end user if they need a persistent
object, that is not returned to the pool automatically.
When you don't longer need the object, be sure to return it
to the pool using \code{poolReturn(object)}.
}
