% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{release}
\alias{release}
\title{Releases an object back to the pool.}
\usage{
release(object)
}
\description{
Should be called by the end user if they previously fetched an
object directly using \code{obj <- pool$fetch()} and are now
done with said object.
}

